requirements.txt
django==3.0.5
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
python-dotenv